gcp/requirements.txt
backports.entry-points-selectable==1.1.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
distlib==0.3.3
filelock==3.4.0
google-api-core==2.2.2
google-auth==2.3.3
google-cloud==0.34.0
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
idna==3.3
platformdirs==2.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.26.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.7
virtualenv==20.10.0
google-api-python-client==2.31.0
google-auth-httplib2==0.1.0
